refactor: EntityGettable 패키지 이동
