+++ src/main/java/org/ktc2/cokaen/wouldyouin/member/application/HostService.java
@@ -2,6 +2,7 @@ package org.ktc2.cokaen.wouldyouin.member.application;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
+import org.ktc2.cokaen.wouldyouin._common.api.EntityGettable;
 import org.ktc2.cokaen.wouldyouin.auth.application.dto.LocalLoginRequest;
 import org.ktc2.cokaen.wouldyouin.member.application.dto.request.create.HostCreateRequest;
 import org.ktc2.cokaen.wouldyouin.member.application.dto.request.edit.HostEditRequest;
